Type update from integer to float for: AUM, FUND_SIZE

+++ nicnbk-data/nicnbk-data-repo-api/src/main/java/kz/nicnbk/repo/model/pe/Firm.java
@@ -17,7 +17,7 @@ public class Firm extends CreateUpdateBaseEntity{
 
     private String firmName;
     private int foundedYear;
-    private int aum;
+    private float aum;
     private int invTeamSize;
     private int opsTeamSize;
     private String locations;
@@ -49,11 +49,11 @@ public class Firm extends CreateUpdateBaseEntity{
     }
 
     @Column(name = "aum")
-    public int getAum() {
+    public float getAum() {
         return aum;
     }
 
-    public void setAum(int aum) {
+    public void setAum(float aum) {
         this.aum = aum;
     }
